fixed so only copies files when they don't already exist

+++ diskIDMain.py
@@ -113,9 +113,10 @@ class kryoMain(QtGui.QMainWindow, diskIDMainGUI.Ui_kryoMain):
           outdir = os.path.join(dest_dir, subdirname.replace(" ","_"))
 
           for trackFile in os.listdir(indirname):
-            self.imageCreationResults.append("copying track " + str(os.path.join(indirname,trackFile)))
-            QtGui.QApplication.processEvents()
-            shutil.copyfile(os.path.join(indirname,trackFile),os.path.join(outTrackDir,trackFile))
+            if not os.path.isfile(os.path.join(outTrackDir,trackFile)):
+              self.imageCreationResults.append("copying track " + str(os.path.join(indirname,trackFile)))
+              QtGui.QApplication.processEvents()
+              shutil.copyfile(os.path.join(indirname,trackFile),os.path.join(outTrackDir,trackFile))
